Extract MovieCard into its own component file and use it in both SeeMoreScreen and MovieSection, Add add to wishlist button in the MovieCard component to add and remove selected movies to/from the Wishlist
